Revert^2 "Add verifier fallback for JVMTI Get/SetLocalVariable"
This reverts commit 99cbfb55fc7ac0f65b1ccdc7076219fcee383b92.

We were incorrectly bounds-checking the register number. We were
treating it as unsigned when it was actual signed. Previously this
wouldn't matter since normally the debug-info won't have any
information for negative slots but by falling back to the verifier
we hit some check failures.

Reason for revert: Fixed underlying issue with bad bounds check.
Bug: 131711256

Change-Id: I0b859ce322f3b23f937b72d735db8f6870c40602