Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review+1
> 
> Not sure it's the right choice, I would have copied the runtime value.

Agreed - why not just pass the version in Runtime? In theory this should be okay as 0 means "unknown version" and I changed the conditions to always check that, but I wouldn't count on the verifier always doing the most conservative thing (zero is never used in prod or well tested) and/or not hard-failing.

Patch-set: 1
Reviewer: Gerrit User 1059815 <1059815@85c56323-6fa9-3386-8a01-6480fb634889>