Update patch set 7
Patch Set 7:

(32 comments)

Quite a lot of changes again, sorry for that.

I'm not happy with asserting ErrorReporter messages; my rationale was that throwing Exceptions instead makes it more difficult/impossible to use streams in certain places (e.g. ApiResolver's constructor), and it would've ballooned the size of this CL even more. But maybe that was better?

Patch-set: 7