Store API policy in oat, check when loading
Hidden and core-platform API access checks are not enabled for all
processes - controlled by command line flags and zygote fork flags.
This information has not been encoded in the oat files, meaning that
if a dex file was AOT-compiled with one API enforcement policy but
ran with another, ART would not reject the oat file as it should.

This patch stores the policy configuration in the oat header's key-
value store and extends OatFileAssistant's getDexOptNeeded logic to
take it into account. Furthermore, libcore's DexFile.getDexOptNeeded
is given an additional argument - target hidden API policy - which
enables the system server to check whether a package needs background
dexopting.

Bug: 79680348
Bug: 79681230
Test: m test-art-{host,target}-gtest-oat_file_assistant_test
Change-Id: I476c3c9f8389bce5aba1a192a5b069eb298670e7