ART: Avoid shared cache lines for JIT code allocations
Dual view JIT exhibits some hygiene issues that causes crashes on
devices with 32-bit kernels.

This change makes JIT code allocations cache aligned. This is based on
guidance in the v7_cohernet_user_range() that says "it is assumed that
the Icache does not read data from the write buffer".

Bug: 132205399

Test: >2000 boot tests on affected device with no zygote crashes.
Test: No crashes running ART JIT benchmarks on go/lem
Test: No failures with Treehugger

Change-Id: I901e2e5c07b9502876b33f572be63ec1dca19cbe