Don't prune the dalvik-cache after flashing a binary
To minimize the use of storage in /system,
I would like to make image which will be saved in /data
including artifacts (vdex/odex/art) of application.
Therefore, I don't need to prune the dalvik-cache during
the first boot just after flashing a binary on device.

When build image for /data, make a file named .stored
for prevent pruning at dalvik-cache.
And then when complete booting, this file is removed.

Test: Flash the binary included dalvik-cache files on device
Change-Id: Id93e82c2ad2ec27f0bac45a6523f67b1bbbe7402