Fix imgdiag crashing for ASLR images
Fixes a bug where the non runtime image was being used to calculate
the size of mirror::Objects, this caused a crash if the image was
ASLR relocated during boot.

This CL fixes imgdiag from crashing and provides a working diff for
app vs zygote. App vs image (for diagnosing shared dirty) is still
incorrect due to ASLR.

Test: test-art-host
Test: adb shell imgdiag --zygote-diff-pid=$(pid zygote64) --image-diff-pid=$(pid system_server) --boot-image=/system/framework/boot.art
Bug: 134614794

Change-Id: I85a5625a60112471778d86a7d00af0b42bc2ef8c