Update patch set 2
Patch Set 2: Code-Review+1

The change looks correct and necessary to me. Thanks!

There is a DCHECK in PushBack that should have caught this in debug builds, so it seems to be unlikely in our environment.

Roland, Lokesh, or Mathieu should look at this as well.

Patch-set: 2
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1