Optimize stack map decoding - interleave varints.
Reorder the layout of consecutive varints.  Store all the 'headers'
which define the varint size first and then store any large values.
The size is unchanged, but it makes the reading from memory faster.

This speeds up CodeInfo by 10%, and maps startup by 0.1%.
Change in size is negligible (the bits mostly just move).

Test: test.py -b --host --64 --optimizing
Change-Id: I16f702389d4ba94a407b1586cae5f82b8ef60e7b