Support using opaque JNI ids
Currently JNI ids (jmethodID &  jfieldID) are created by simply
casting the corresponding ART structure pointer. This is great for
simplicity but means we are prevented from performing operations that
could change these pointer values. To support these use-cases add
support for loading the runtime with a layer of indirection between
these ids and the internal art data types.

Currently the JNI id type can be toggled only by passing the new
'-Xopaque-jni-ids:{true,false}' flag during startup.

This adds runs with and without opaque-jni-ids to the art-jit test
configuration. I put it in this configuration since that is where we
test --debuggable which we expect to use opaque-jni-ids with in the
future.

Test: ./test.py --host --index-ids --pointer-ids
Test: ./test/testrunnner/run_build_test_target.py art-jit
Bug: 134162467
Change-Id: Id8c8cb9a5b8ff18dc2f40892fae2d344a7214f44