Update patch set 1
Patch Set 1:

> How does this help determinism? Eventually we will want to support app finalizers in app images if we enable freeing app image objects (e.g. by copying the app image to the heap).

I linked to the determinism bug because this is a follow-up to https://android-review.googlesource.com/983105 that goes toward addressing a FIXME introduced in that CL.

We do not support finalizable object allocation from a transaction. The only finalizer references we construct for app image compilation are for the ClassLoader that's used only for a runtime sanity check and then replaced in all objects, see "AppImage:UpdateClassLoaders" section in ClassLinker::AddImageSpace(). The ClassLoader (and subobjects like dalvik.system.DexFile) does not actually need to be finalized.

Patch-set: 1
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>