Introduce a Testing Runtime APEX including ART gtests.
Define a new flavor of the Runtime APEX, having the same contents
as the Debug Runtime APEX plus ART gtests that have dependencies
on Runtime APEX internal libraries.

Previously, when these ART gtests were built and installed in
the data partition, they would depend on "legacy" ART internal
libraries being installed in the system partition (not the ones
from the Runtime APEX). As we plan to remove these ART libraries
from the system partition, we needed a way to be able to build
and install device ART gtests without triggering the installation
of these legacy ART internal libraries.

This Testing Runtime APEX is meant to be used for ART testing only,
and should never be included in any product.

Test: m com.android.runtime.testing
Test: Run gtests on device in chroot environment
Bug: 129534335
Change-Id: I8d70bdcf7c6253dd5bc8786a702f130d444c3cb6