WIP: Add ability to switch to index ids late - SLIM.
In order to support some debugging features we need to have JNI code
mostly running with jmethodIDs and jfieldIDs as opaque indexes
disconnected from any underlying art data-structure pointers. For
performance though we want to continue to use data-structure pointers
as these IDs when debugging is not needed. To support both
possibilities this adds support for changing from a 'swapablePointer'
to either 'pointer' or 'indicies' regime at runtime.

The SwapablePointer regime still uses pointers as the ids but marks
any classes using which have ids retrieved and always returns pointers
for those classes. Some care is required about where & when the mode
is changed in order to avoid races.

Test: ./test.py --host
Bug: 134162467
Change-Id: Ib1d05501be4eb3931763ac0f3bbbc6943b4ef7ef