Update patch set 1
Patch Set 1:

> LGTM, but I don't understand enough of the big picture here. It makes sense to me that there's a race here, and this is needed, but I can't confirm that. [...]

Just imagine that one thread constructs and interns a String, stores it (relaxed) in .bss/DexCache, and another thread loads it from there (relaxed) and fetches the length. There's a chance it's going to see 0 because the memory was zeroed before constructing the String, right? (Though not having the memory in L1 cache will probably cause the loading thread to see the newly stored value, I guess. But we should not rely on that.)

Some time ago we changed the DexCache stores to be release operations to be safe, see also b/32075261 . But I see we missed SetResolvedString() and SetResolvedMethodType() which still use relaxed stores.

Patch-set: 1
Reviewer: Gerrit User 1018108 <1018108@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0