Handle region space objects while sweeping live stack
Currently we donot remove the region space objects
from the object list before processing for LOS.
And since we dont check if LOS contains the object
to be deleted, we attempt LOS::Free on region space
objects resulting in a crash.

Ignoring region space objects fixes this.
All the other objects must belong to LOS.

Test: ./test.py --host

Change-Id: Ieb78876182bc057c28d852d4b96263b7e9a44616
Signed-off-by: Vinay Kompella <vinay.kompella@intel.com>