ARM/ARM64: Improve frame entry/exit codegen.
On ARM64, use STP pre-index for the method and the lowest
spilled core register for method entry if there's no gap or
FP spills in between. On exit, use LDP post-index to restore
in this case, clobbering a scratch register. Thus, we save
one instruction for both entry end exit for such methods and
the performance should be the same or better.

On ARM, use a single PUSH/POP for method entry and core
spills if the gap between them is 2 words or less and and we
have one or no FP spill, spill args as filler if needed. On
exit, load the FP spill if any and do a single POP for core
registers and return in this situation, clobbering as many
registers from r2-r4 as needed; these caller-save registers
are not used to pass return values. If we cannot do this
because of FP spills but the gap between the method and FP
spills is 2 words or less, we adjust SP and save the method
in one PUSH after spilling; there is no similar handling
for method exit as the method does not need to be restored.
This may improve or degrade performance a bit depending on
the particular situation; in the worst case we PUSH/POP
three additional registers as a cost for smaller code size.

aosp_taimen-userdebug prebuils:
 - before:
   arm/boot*.oat: 19147484
   arm64/boot*.oat: 22558344
   oat/arm/services.odex: 21922256
 - after:
   arm/boot*.oat: 19105436 (-41KiB, -0.2%)
   arm64/boot*.oat: 22549392 (-9KiB, -0.04%)
   oat/arm/services.odex: 21914032 (-8KiB, -0.03%)

Test: aosp_taimen-userdebug boots.
Test: run-gtests.sh
Test: testrunner.py --target --optimizing
Bug: 136144107
Change-Id: Id36c67b4e735418fb18bcd3269b72b25695fbaa2