Install appcompat.sh for use by make rules
The make build for apps (including prebuilts) hard codes a reference to
the art/tools/veridex/appcompat.sh file which causes a build breakage
when unbundling the runtime module, i.e. when the art repository is not
present in the local build tree.

This change installs the appcompat.sh into $(HOST_OUT)/bin/appcompat/
instead. That allows the unbundled build to provide a prebuilt that
installs the script into the same location.

Bug: 134379140
Test: m droid
Change-Id: If1a4ea1f2798b6fb27b4feb76f878cba8939aabc