WIP: resize arrays JVMTI extension
As a prototype for more general object replacement functionality add a
new JVMTI extension that allows one to change the size of arrays. This
extension is 'com.android.art.heap.change_array_size'. This extension
atomically replaces every reference (both strong and weak, global and
local) with a newly allocated array with the same contents but a
different length.

Test: ./test.py --host
Bug: TBD

Change-Id: I92a0beabb02e0c92c8c8f9639836014ff1266878