Update patch set 2
Patch Set 2:

(3 comments)

D8 1.6.17-dev includes an optimization that removes const-instructions when another non-const value is known to have the same value at runtime.

Example:

  boolean b = unknown();
  if (b) {
    int x = 1;
    // In the following, since `b` is known to be 1, `x` is replaced by `b`,
    // such that the const instruction can be eliminated.
    consume(x);
  }

This affects some ART tests. In particular, several tests contain the following snippet:

  if (doThrow) { throw new Error(); }

Since `doThrow` is then known to be 0 in the remainder of the code, any const instruction that loads 0 will be eliminated. A simple fix is to simply remove the if-statement from the tests if it is not strictly required for the tests.

Patch-set: 2
Reviewer: Gerrit User 1038443 <1038443@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1327199 <1327199@85c56323-6fa9-3386-8a01-6480fb634889>