Update patch set 1
Patch Set 1:

I meant on a CodegenCommonArchProperties structure that contains the srcs from all of the enabled architectures, although it may be cleaner to store in a []string, filter it, and then put it in a struct for AppendProperties.

Also, these are strings, not paths, so it should be android.FirstUniqueStrings.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>