Update patch set 3
Patch Set 3: Code-Review+1

I'm fine with this.

But I have an alternative proposal: Check in the test-dumps as prebuilts and get rid of the build rules for generating them. That way we don't need to maintain the complex build rules for running ART as part of the build, we don't need any special build support for ahat, and ahat no longer has any dependency on ART.

It means we will not get notified proactively if a change in ART or elsewhere in the platform (such as changes to proguard mappings) breaks ahat tests, but I think that's a reasonable tradeoff. The ahat tests have only minimal value in testing ART, and any other failures in the tests due to ART changes would need fixes in ahat itself. If needed we could generate new prebuilts for testing ahat against each major platform release of ART.

Thoughts?

Patch-set: 3
Reviewer: Gerrit User 1057373 <1057373@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1