Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> The "sync" bit about the two files is really only about keeping in sync what ART binaries need. Because they don't need neuralnetwrorks, you shouldn't need this change.

Well, as the author of that comment I can say that I actually meant them to be in sync as a matter of course. The rationale is 1) the less differences there are the easier it is to reason about things, and 2) with autogenerated files we'll probably get these links anyway so there's less potential for trouble if they are here from the start.

That said, the discussion here shows this approach doesn't quite work, so I'm not pushing for resurrecting this CL. Hopefully we won't be living with these files for much longer. :\

Patch-set: 9