Reenable LUCI notifications on the ART Buildbot AOSP builders.
The clean steps on the AOSP builders (aosp-builder-cc and
aosp-builder-cms) used to call "make clean", which sometimes failed
with this error:

  Failed to remove file "out/.module_paths":
  unlinkat out/.module_paths: directory not empty

As a result the LUCI notitications for these builders were disabled.

The clean steps are now using "rm -rf out" instead of "make clean",
which presubmably addressed the previous issue. Reenable the LUCI
notifications for aosp-builder-cc and aosp-builder-cms.

Test: Rely on the ART Buildbot
      (https://ci.chromium.org/p/art/builders/ci/aosp-builder-cc
      and https://ci.chromium.org/p/art/builders/ci/aosp-builder-cms)
Bug: 131573205
Bug: 139655167

Change-Id: I38f31d2ccf1f337698aaf463d90efc593c666976