Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> Why is this an art test? Specifically, shouldn't it be a libcore JUnit test (that forms part of CTS)? From what I can see, it doesn't rely on any ART specifics?

I'm open to moving it. If we put it there, would we be able to refer to the stuff in external/ rather then replicating it?

My assumption was that this is most likely to be broken by ART changes that try to directly use machine instructions for some of these functions.

Patch-set: 5
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>