Add spin loop to mutex, overhaul monitor.
DO NOT SUBMIT. The new test will fail since it currently generates
unexpected timing output.

Since Linux context switch overhead is typically larger than a
microsecond, this may greatly reduce the overhead of waiting for a
mutex that is only briefly held by another thread. Rather than going
to sleep and having to be woken back up again, at a cost of several
microseconds, we just spin, hopefully for much less than microsecond,
until the mutex becomes available. It does waste some CPU cycles
when spinning fails, either because the lock is held too long, or
we are being scheduled against the thread holding the lock. But we
expect those to be unlikely.

We test the lock only a few times, with pauses in between. It's
unclear that's beneficial; we should perhaps just loop reading
the variable.

Add a test that mutual exclusion works, which can also be run as
lock microbenchmark.

The old monitor implementation did not benefit much from this, because
it used mutex only as a low-level lock to protect the monitor data
structure. Instead use monitor_lock_ as the actual lock protecting the
data structure, i.e. hold onto it while the monitor is fatlocked.
Among other things, this requires that the monitor_lock_ always
be acquired by, or explicitly on behalf of, the thread holding the
monitor.

This in turn makes it really hard to deflate a monitor held by another
thread. Just stop doing that, since it was unclear whether that's
actually beneficial.

Very tentatively, this appears to get us about a decimal order of
magnitude in inflated contended locking performance.

Bug: 111835365
Test: Successfully built and ran 2029-contended-monitors. More coming.
Change-Id: Iff7457fff59efcb24e25d35a4ef71b67b8a9082a