Add a brief spin loop to mutex waiting operations
DO NOT SUBMIT. The new test will fail since it currently generates
unexpected timing output.

Since Linux context switch overhead is typically larger than a
microsecond, this may greatly reduce the overhead of waiting for a
mutex that is only briefly held by another thread. Rather than going
to sleep and having to be woken back up again, at a cost of several
microseconds, we just spin, hopefully for much less than microsecond,
until the mutex becomes available. It does waste some CPU cycles
when spinning fails, either because the lock is held too long, or
we are bing scheduled against the thread holding the lock. But we
expect those to be unlikely.

We test the lock only a few times, with pauses in between. It's
unclear that's beneficial; we should perhaps just loop reading
the variable.

Add a test that mutual exclusion works, which can also be run as
lock microbenchmark.

Based on the benchmark, the mutex.cc change improves inflated
(and contended) locking times by about 10%. But there is more work to
be done here.

Bug: 111835365
Test: Booted AOSP
Change-Id: Iff7457fff59efcb24e25d35a4ef71b67b8a9082a