Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> How are we evaluating this change quantitatively?

Timely question. I'm working on a microbenchmark to understand this better. There seems to be no meaningful impact on go/lem benchmarks in either direction. which is surprising to me, and suggests that either we need more benchmarks, or that my understanding of what's going on is faulty.

If you have other suggestions on how to evaluate this, that would be great. 

We've been seeing some scalability issues with benchmarks that might be explained by the way we currently deal with lock contention. So far my immediate goal is to understand the issues better. For very well-written apps, this probably shouldn't matter either way. But I would be surprised if there isn't a bunch of code for which it does.

AFAICT, once a Java monitor is inflated, we currently immediately wait on a futex whenever we encounter contention. And AFAIK, the futex wait currently results in a full context switch unless the futex condition is already satisfied when we enter the kernel? That seems clearly suboptimal, and my intuition is we should be able to do substantially better. As we discussed earlier, there may be ways to do better than this CL. But it would be surprising to me if this CL doesn't help.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0