hiddenapi: add 'test-api' flag into dex files.
Only CTS tests should have access to @TestApi signatures. But now all of them are greylisted, hence can be accessed by anyone. This should be changed so that grey/black/whitelist properties would be orthogonal to test-api.

The first step to fix this is to introduce extra flag and embed it into dex files. In the future access decisions will be made based on that flag.

Bug: 133832325
Test: make -j32 test-art-host-gtest-hiddenapi_test
Change-Id: I0db8938391d6aefe5e9a3bc51ec99b948ae41f6d