Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
...
> > 
> > Any idea how common it is to wait for a lock on a non-runnable process?
> 
> No idea, but I think I know how to find out. Each time we spin, record the before-spin and after-spin timestamps, along with the owning thread's TID, into a small per-thread buffer, flushing the buffer occasionally. (We want to buffer the timestamps because writing a trace event for each contention would be too expensive, I think.) We can then intersect those spin intervals with thread scheduling using DCTV and figure out how much time we spend spinning unproductively. If we collect a long trace with this instrumentation added (or run DOU or something) we can get a sense of how productive the spinning is under a typical workload.

That seems like it would work, but sounds pretty elaborate. We could presumably also get a good idea by just keeping success and failure counters in the spin code. That wouldn't get us quite the same data, but I wonder whether it wouldn't be closer to what we actually want. It would include time spent spinning for a running thread holding a lock for a long time. And in some racing cases, we could be successful even if the owning thread were initially descheduled. But it seems like this data would be a better reflection of our actual success rate?

We may eventually want to keep per-mutex success and failure counters anyway, so we can make the spinning adaptive, and avoid doing so for mutexes that are normally held for a long time.

I'd still like to get the microbenchmarks first, to make sure this behaves roughly as I expect.

Patch-set: 1
Reviewer: Gerrit User 1042828 <1042828@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0