Update patch set 2
Patch Set 2: Verified+1

I'm not really used to this code, but this is required on some of our devices.
You added proper kernel version checks in the memfd_create function but don't check for the errno, so I think this is appropriate.
Feel free to rewrite the commit message so that it makes sense to you.

Patch-set: 2
Reviewer: Gerrit User 1081223 <1081223@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1