Consolidate profile methods
Remove unnecessary profile APIs or the ones that were added only for tests.
This makes the flow of data much easier to understand (as we have to follow fewer entry points when adding data to the profile). It will make it easier to restructure the profile format with new data.

Test: m test-art-host-gtest
Change-Id: I09ea91229278877d16b12af7a0b356ccceb4b520