Restructure the profile tests
Restructure the test to manipulate the profile using the same methods as the
profile saver (rather than using index based addition).

This will help with cleaning up the profile interface in preparation for a new format.

To support the restructuring new classes were added to help create and manage
fake dex files which only have the location, checksum and number of methods set.

Bug: 139884006
Test: m test-art-host
Change-Id: I9bf86b2244c01ffa9422df0fd5ee23a2a7ff8593