Process method flags uniformly in the profile
We used to have two main ways to add methods in profiles: one which would assume a method is hot and another where will had to specify the method flags explicitly. This created some implicit assumptions when adding new methods that made it hard to follow which methods sets what.

This CL unify the way flags are processed removing any assumptions of flags.

Bug: 139884006
Test: m test-art-host

Change-Id: Ib1dcfdd4a5220507b89e3e0947252cef4fa2eb4c